Create Problem D. Stages.cpp
RU CSE Practice Contest Nov 11, 2022
-
Problem D. Stages
Time limit 1000 ms
Mem limit 262144 kB
Natasha is going to fly to Mars. She needs to build a rocket, which consists of several stages in
some order. Each of the stages is defined by a lowercase Latin letter. This way, the rocket can be
described by the string — concatenation of letters, which correspond to the stages.
There are stages available. The rocket must contain exactly of them. Stages in the rocket
should be ordered by their weight. So, after the stage with some letter can go only stage with a
letter, which is at least two positions after in the alphabet (skipping one letter in between, or
even more). For example, after letter 'c' can't go letters 'a', 'b', 'c' and 'd', but can go letters 'e',
'f', ..., 'z'.
For the rocket to fly as far as possible, its weight should be minimal. The weight of the rocket is
equal to the sum of the weights of its stages. The weight of the stage is the number of its letter in
the alphabet. For example, the stage 'a 'weighs one ton,' b 'weighs two tons, and' z' — tons.
Build the rocket with the minimal weight or determine, that it is impossible to build a rocket at
all. Each stage can be used at most once.
Input
The first line of input contains two integers — and ( ) – the number of
available stages and the number of stages to use in the rocket.
The second line contains string , which consists of exactly lowercase Latin letters. Each letter
defines a new stage, which can be used to build the rocket. Each stage can be used at most once.
Output
Print a single integer — the minimal total weight of the rocket or -1, if it is impossible to build
the rocket at all.
Sample 1
Input Output
5 3
xyabd
29
Sample 2
Input Output
n k
26
n k 1 ≤ k ≤ n ≤ 50
s n
RU CSE Practice Contest Nov 11, 2022
-
Input Output
7 4
problem
34
Sample 3
Input Output
2 2
ab
-1
Sample 4
Input Output
12 1
abaabbaaabbb
1
Note
In the first example, the following rockets satisfy the condition:
"adx" (weight is );
"ady" (weight is );
"bdx" (weight is );
"bdy" (weight is ).
Rocket "adx" has the minimal weight, so the answer is .
In the second example, target rocket is "belo". Its weight is .
In the third example, , so the rocket must have both stages: 'a' and 'b'. This rocket
doesn't satisfy the condition, because these letters are adjacent in the alphabet. Answer is -1.
1 + 4 + 24 = 29
1 + 4 + 25 = 30
2 + 4 + 24 = 30
2 + 4 + 25 = 31
29
2 + 5 + 12 + 15 = 34
n = k = 2